bug 17698 - CD device state STATE_LOOKINGUPCD isn't sticking (r=kreeger)
git-svn-id: http://publicsvn.songbirdnest.com/client/trunk@14655 5aad4ff5-4b58-4159-ba3f-0e6328df4cad

+++ components/devices/cd/src/sbCDDevice.h
@@ -256,7 +256,7 @@ public:
   /**
    * CD device specific status.
    */
-  static const int STATE_LOOKINGUPCD = STATE_USER + 1;
+  static const unsigned long STATE_LOOKINGUPCD = STATE_USER + 1;
 
 private:
   /**
